refactor(package): extract vcs check to a separate module
